+++ app/src/main/java/dev/arkbuilders/rate/domain/model/CurrencyRate.kt
@@ -1,5 +1,7 @@
 package dev.arkbuilders.rate.domain.model
 
+import java.math.BigDecimal
+
 enum class CurrencyType {
     FIAT,
     CRYPTO,
@@ -8,5 +10,5 @@ enum class CurrencyType {
 data class CurrencyRate(
     val type: CurrencyType,
     val code: CurrencyCode,
-    val rate: Double,
+    val rate: BigDecimal,
 )
